+++ public/components/layer_config/cluster_config/data_source_section.tsx
@@ -3,16 +3,17 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import React from 'react';
+import React, { useEffect } from 'react';
 import {
   EuiFormRow,
+  EuiLoadingSpinner,
 } from '@elastic/eui';
 import { IndexPattern } from '../../../../../../src/plugins/data/public';
 import { useOpenSearchDashboards } from '../../../../../../src/plugins/opensearch_dashboards_react/public';
 import { MapServices } from '../../../types';
 import { i18n } from '@osd/i18n';
 import { CanUpdateMapType } from './cluster_layer_source';
-import { useEffect } from 'react';
+import { useIndexPatternFilter } from '../../../utils/use_index_pattern_filter';
 
 interface Props {
   indexPattern: IndexPattern | null | undefined;
@@ -35,13 +36,22 @@ export const DataSourceSection = ({ indexPattern, setIndexPattern, setCanUpdateM
       },
     },
   } = useOpenSearchDashboards<MapServices>();
-  
+  const { filter: indexPatternFilter, loading: indexPatternFilterLoading } = useIndexPatternFilter(savedObjectsClient);
+
   useEffect(() => {
     setCanUpdateMap((prev) => ({
       ...prev,
       index: !!indexPattern,
     }));
   }, [indexPattern]);
+
+  if (indexPatternFilterLoading) {
+    return (
+      <EuiFormRow label="Index pattern" fullWidth>
+        <EuiLoadingSpinner size="m" />
+      </EuiFormRow>
+    );
+  }
 
   return (
     <EuiFormRow
@@ -64,6 +74,7 @@ export const DataSourceSection = ({ indexPattern, setIndexPattern, setCanUpdateM
         data-test-subj={'indexPatternSelect'}
         fullWidth={true}
         isInvalid={!indexPattern}
+        indexPatternFilter={indexPatternFilter}
       />
     </EuiFormRow>
   );